prepare for 64 bit

+++ kernel/memory/pagemgr.h
@@ -24,7 +24,7 @@
 
 #define DYNAMIC_START   MB16    // start dynamic allocation of pages at phys addr 16MB
 
-void                setup_pagemgr(uint32_t available_memory);
+void                setup_pagemgr(unsigned long available_memory);
 pagetable_entry_t   get_free_page(int flags);
 void                release_page(void *page);
